Cypress installed
