+++ client/cypress/e2e/submit-input.cy.js
@@ -1,0 +1,15 @@
+import { dataGenerator } from '../support/custom/generate-input';
+import { startSession } from '../support/custom/session-create';
+describe('deepEqual', () => {
+it('create and encrypt user input', () => {
+    // Create csv
+  const input = dataGenerator();
+  console.log(input);
+    // input id, code, industry, size
+    // Drop the csv
+    // Click box
+    // Click submit
+
+    // console.log(startSession());
+  });
+});
