+++ client/package.json
@@ -40,6 +40,7 @@
     "@types/react": "^18.0.27",
     "@types/react-dom": "^18.0.10",
     "@vitejs/plugin-react-swc": "^3.0.0",
+    "cypress": "^12.9.0",
     "jest": "^29.5.0",
     "prettier": "^2.8.4",
     "ts-jest": "^29.0.5",
